organize things a bit better
